back-end/src/core/typedefs/result_future.ts
import { Failure } from "../errors/failure";

export type Either<L, R> = Left<L> | Right<R>;

export type ResultFuture<T> = Promise<Either<Failure, T>>;

export class Left<L> {
  readonly value: L;

  constructor(value: L) {
    this.value = value;
  }

  isLeft(): this is Left<L> {
    return true;
  }

  isRight(): this is Right<any> {
    return false;
  }
}

export class Right<R> {
  readonly value: R;

  constructor(value: R) {
    this.value = value;
  }

  isLeft(): this is Left<any> {
    return false;
  }

  isRight(): this is Right<R> {
    return true;
  }
}
